Room Options Implemented / Bug Fix
*) The ability to read in building files and their associated room file
into a combo box.
*) Each campus has an associated folder to place all known building and
room files.
*) It is required to have an index.txt file stating the building name
followed by the associated room file for that building. (Could be used
for multiple language support)
*) Fixed issues when selecting a different campus and CMU
buildings/rooms were still loaded.
*) Assumption is made (for simplicity sake) that directions to two
different regions are within the same campus. When a user selects a
different campus, all combo boxes are flushed and files for the
associated buildings/rooms are loaded with respect to the campus.
*) A global record called "CampusBuildings" contains all building
information and rooms with the selected campus.
(populate_getDirections() and [combobox]OnChange() must be called for
"CampusBuildings" to be populated)